chore: sort various names, automate ESLint sorting
Installed the Perfectionist ESLint plugin and configured it to sort
imports, JSX props and many other things alphabetically. And sorted
everything, of course, immediately.

It is very far from ideal and I don't know if what it brings is better
than manual sorting, but I'll try using it for some time. At least
import statement sorting is great. This is why I decided to use it in
the first place.