fix: keyboard navigation issue
When there was only the "Personal" section and the focus was on it,
the user couldn't get to the control buttons. Also, if there was more
than one section, but the last one was opened, the user couldn't get to
the navbar control buttons by simply pressing `Tab`.

These issues were fixed in this commit by making it possible to navigate
between all the navbar children by arrow keys.