feat: add most sect., update input change handling
Added all sections but "Certifications". "Experience" was implemented
completely, but others are yet to be finished.

As for "input change handling", I moved from verbose and repetitive
`onChange` event handlers to adding `handleInputChange` functions to all
components and putting them as event handlers which reduced the
repetetiveness of the code.