feat: improve preview adaptivity, add its header
Right now the preview opens by default as soon as the app loads. It's
more convenient while I'm working on the modal.

There's a problem with performance right now because of `Font.register`.
I guess there's no way around it but switching from native dialogs to
conditional rendering. Otherwise, `Font.register` is called every time
the browser executes the `Preview` component. What we need is that it's
called only when the modal is rendered.