refactor: improve PDF rendering performance
The bottleneck that caused up to 10 seconds of rendering (on the main
thread!) was fonts, and I did two things to fix this,

- Switched to Latin subsets.
- Switched to TTF from WOFF.

The former is obvious, and the latter improved performance because WOFF
is a compressed format, and the decompression process takes quite a lot
of time for the `react-pdf` library's engine.

Oh, and I decided to render `<PDFViewer>` conditionally instead of
simply keeping it inside `<Popup>`. Even when the modal isn't on the
screen, it's still executed. Without conditional rendering,
`<PDFViewer>`, as `<Popup>`'s child, was also executed even when it
wasn't displayed. That decreased performance quite significantly. Not
anymore.