refactor: make returned fns pure in `useAppState`
In the process, I had to merge states `activeSectionIds` and
`openedSectionId` into one state `sectionsState`, which in turn
necessitated refactoring the `useEffect`s.

Also, there's no need in the utility function
`getArrayWithoutDuplicates` and, because it'd been created specifically
for `useAppState`, I deleted it.

And one more thing. This commit brings many smaller refactors that
change internal implementation of returned functions. Mostly, its for
better performance. But it doesn't change anything for the hook's users.